requirements.txt
pymongo>=4.0.0
pdf2image>=3.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
openai>=1.12.0
google-cloud-vision>=3.0.0